+++ app/src/main/java/com/battlelancer/seriesguide/shows/episodes/EpisodeDetailsFragment.kt
@@ -65,6 +65,7 @@ import com.battlelancer.seriesguide.util.Utils
 import com.battlelancer.seriesguide.util.ViewTools
 import com.battlelancer.seriesguide.util.copyTextToClipboardOnLongClick
 import com.battlelancer.seriesguide.util.safeShow
+import com.battlelancer.seriesguide.util.startActivityWithAnimation
 import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
 import kotlinx.coroutines.Job
@@ -447,13 +448,13 @@ class EpisodeDetailsFragment : Fragment(), EpisodeActionsContract {
 
         // episode image
         val imagePath = episode.image
-        binding.containerImage.setOnClickListener { v: View? ->
+        binding.containerImage.setOnClickListener { v: View ->
             val intent = intent(
                 requireContext(),
                 ImageTools.buildEpisodeImageUrl(imagePath, requireContext()),
                 ImageTools.buildEpisodeImageUrl(imagePath, requireContext(), originalSize = true)
             )
-            Utils.startActivityWithAnimation(requireActivity(), intent, v)
+            requireActivity().startActivityWithAnimation(intent, v)
         }
         loadImage(imagePath, hideDetails)
 
@@ -594,12 +595,12 @@ class EpisodeDetailsFragment : Fragment(), EpisodeActionsContract {
         }
 
         // Trakt comments
-        bindingButtons.buttonEpisodeComments.setOnClickListener { v: View? ->
+        bindingButtons.buttonEpisodeComments.setOnClickListener { v: View ->
             val episodeId = episodeId
             if (episodeId > 0) {
                 val intent =
                     TraktCommentsActivity.intentEpisode(requireContext(), episodeTitle, episodeId)
-                Utils.startActivityWithAnimation(requireActivity(), intent, v)
+                requireActivity().startActivityWithAnimation(intent, v)
             }
         }
     }